+++ lib/human_player.ex
@@ -1,10 +1,9 @@
 defmodule HumanPlayer do
-  defstruct [:move]
-  def move(grid) do
-    CLI.chosen_move(grid,"X", "O")
-  end
+  defstruct [:mark]
 end
 
 defimpl Player, for: HumanPlayer do
-  def get_move(human_player), do: human_player.move
+  def get_move(_human_player, grid) do
+    CLI.chosen_move(grid, "X", "O")
+  end
 end
